Optimize the output of zips with directory entries
Using DEFLATE (default) bundles a header-section that needs to be decompressed in order to know the true content. Even for an empty byte array this creates 18 bytes of junk. Using STORED instead we set the size to 0 manually and it stays that way.

For the same reason we will always use STORED for empty files in ZIP output as well.